started working on redesign
