+++ papiea-engine/package.json
@@ -33,6 +33,7 @@
     "@types/btoa": "^1.2.3",
     "@types/deep-equal": "^1.0.1",
     "@types/jaeger-client": "^3.15.4",
+    "@types/object-hash": "^1.3.4",
     "ajv": "^7.1.0",
     "atob": "^2.1.2",
     "axios": "^0.18.1",
@@ -49,6 +50,7 @@
     "mnemonist": "^0.30.0",
     "mongodb": "^3.1.10",
     "morgan": "^1.9.1",
+    "object-hash": "^2.1.1",
     "opentracing": "^0.14.5",
     "papiea-backend-utils": "file:../papiea-backend-utils",
     "papiea-core": "file:../papiea-core",
